third_party/blink/web_tests/fast/multicol/vertical-lr/balancing/balance-unbreakable.html

<!DOCTYPE html>
<script src="../../../../resources/js-test.js"></script>
<p>There should be a <em>square</em> with some stuff inside below.</p>
<div id="console"></div>
<div id="mc" style="-webkit-writing-mode:vertical-lr;-webkit-text-orientation:sideways-right;text-rendering:optimizeLegibility; -webkit-columns:3; columns:3; height:7em; orphans:1; widows:1; background:olive;">
    <div style="line-height:4em;">line</div>
    <div style="line-height:8em;">line</div>
    <img style="display:block; width:2em;" src="../../resources/blimp.png">
    <div style="line-height:4em;">line</div>
    <div style="line-height:10em;">line</div>
    <div style="line-height:2em;">line</div>
    <div style="line-height:6em;">line</div>
</div>
<div style="width:14em; height:7em; background:olive;"></div>
<script>
    description("Tests that an image at a column boundary is recorded properly for balancing");
    shouldBe("mc.offsetWidth", "224");
</script>
