third_party/blink/web_tests/platform/generic/fast/multicol/vertical-lr/balancing/balance-unbreakable-expected.txt

There should be a square with some stuff inside below.

Tests that an image at a column boundary is recorded properly for balancing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mc.offsetWidth is 224
PASS successfullyParsed is true

TEST COMPLETE

line
line
line
line
line
line
